migrated from json to pickle
reason is that json does not allow me to use integer values for dictionary keys. this feature is a must have requirement for my models to work as intended. though, right now the data in the database is non-readable by human, it just works and i am a happy man.